+++ src/sagemaker/session.py
@@ -107,8 +107,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
         """
         self._default_bucket = None
         self._default_bucket_name_override = default_bucket
-
-        # currently is used for local_code in local mode
+        self.s3_resource = None
+        self.s3_client = None
         self.config = None
 
         self._initialize(
@@ -199,7 +199,10 @@ class Session(object):  # pylint: disable=too-many-public-methods
             key_suffix = name
 
         bucket = bucket or self.default_bucket()
-        s3 = self.boto_session.resource("s3")
+        if self.s3_resource is None:
+            s3 = self.boto_session.resource("s3", region_name=self.boto_region_name)
+        else:
+            s3 = self.s3_resource
 
         for local_path, s3_key in files:
             s3.Object(bucket, s3_key).upload_file(local_path, ExtraArgs=extra_args)
@@ -227,7 +230,11 @@ class Session(object):  # pylint: disable=too-many-public-methods
             str: The S3 URI of the uploaded file.
                 The URI format is: ``s3://{bucket name}/{key}``.
         """
-        s3 = self.boto_session.resource("s3")
+        if self.s3_resource is None:
+            s3 = self.boto_session.resource("s3", region_name=self.boto_region_name)
+        else:
+            s3 = self.s3_resource
+
         s3_object = s3.Object(bucket_name=bucket, key=key)
 
         if kms_key is not None:
@@ -254,7 +261,10 @@ class Session(object):  # pylint: disable=too-many-public-methods
 
         """
         # Initialize the S3 client.
-        s3 = self.boto_session.client("s3")
+        if self.s3_client is None:
+            s3 = self.boto_session.client("s3", region_name=self.boto_region_name)
+        else:
+            s3 = self.s3_client
 
         # Initialize the variables used to loop through the contents of the S3 bucket.
         keys = []
@@ -299,7 +309,10 @@ class Session(object):  # pylint: disable=too-many-public-methods
             str: The body of the s3 file as a string.
 
         """
-        s3 = self.boto_session.client("s3")
+        if self.s3_client is None:
+            s3 = self.boto_session.client("s3", region_name=self.boto_region_name)
+        else:
+            s3 = self.s3_client
 
         # Explicitly passing a None kms_key to boto3 throws a validation error.
         s3_object = s3.get_object(Bucket=bucket, Key=key_prefix)
@@ -317,7 +330,10 @@ class Session(object):  # pylint: disable=too-many-public-methods
             [str]: The list of files at the S3 path.
 
         """
-        s3 = self.boto_session.resource("s3")
+        if self.s3_resource is None:
+            s3 = self.boto_session.resource("s3", region_name=self.boto_region_name)
+        else:
+            s3 = self.s3_resource
 
         s3_bucket = s3.Bucket(name=bucket)
         s3_objects = s3_bucket.objects.filter(Prefix=key_prefix).all()
@@ -330,6 +346,7 @@ class Session(object):  # pylint: disable=too-many-public-methods
             str: The name of the default bucket, which is of the form:
                 ``sagemaker-{region}-{AWS account ID}``.
         """
+
         if self._default_bucket:
             return self._default_bucket
 
@@ -364,10 +381,14 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 already being created, no exception is raised.
 
         """
-        bucket = self.boto_session.resource("s3", region_name=region).Bucket(name=bucket_name)
+        if self.s3_resource is None:
+            s3 = self.boto_session.resource("s3", region_name=region)
+        else:
+            s3 = self.s3_resource
+
+        bucket = s3.Bucket(name=bucket_name)
         if bucket.creation_date is None:
             try:
-                s3 = self.boto_session.resource("s3", region_name=region)
                 if region == "us-east-1":
                     # 'us-east-1' cannot be specified because it is the default region:
                     # https://github.com/boto/boto3/issues/125